+++ src/render/templates.rs
@@ -49,6 +49,7 @@ impl TemplateRenderer {
         context.insert("prev_page", &ctx.prev_page);
         context.insert("next_page", &ctx.next_page);
         context.insert("color_mode", &ctx.color_mode);
+        context.insert("js_cachebust", &ctx.js_cachebust);
 
         self.tera
             .render("layout.html", &context)
@@ -79,4 +80,5 @@ pub struct PageContext {
     pub prev_page: Option<PageLink>,
     pub next_page: Option<PageLink>,
     pub color_mode: String,
+    pub js_cachebust: String,
 }